moved original utilities.
