+++ systems/fate/client/consequences.vue
@@ -208,12 +208,12 @@
 
 <script>
     //------------------------------------------------------------------------------------------------------------------
-    
+
     import _ from 'lodash';
 
     // Pull in the shortID utility
-    import { shortID } from '../../../server/utilities';
-    
+    import { shortID } from '../../../server/utils/misc';
+
     //------------------------------------------------------------------------------------------------------------------
 
     export default {
