+++ systems/fate/client/stunts.vue
@@ -132,7 +132,7 @@
     import marked from 'marked';
 
     // Pull in the shortID utility
-    import { shortID } from '../../../server/utilities';
+    import { shortID } from '../../../server/utils/misc';
 
     //------------------------------------------------------------------------------------------------------------------
